chathist/requirements/requirements.txt
hydra-core==1.3.2
numpy<2
pandas==2.2.3
tiktoken==0.9.0
torch==2.7.0
transformers==4.51.3
